TODO FUNCIONA

+++ postinstall.js
@@ -8,7 +8,9 @@ function copyFolderSync(from, to) {
         const fromPath = path.join(from, element);
         const toPath = path.join(to, element);
         if (fs.lstatSync(fromPath).isFile()) {
-            fs.copyFileSync(fromPath, toPath);
+            if (!fs.existsSync(toPath)) {
+                fs.copyFileSync(fromPath, toPath);
+            }
         } else {
             copyFolderSync(fromPath, toPath);
         }
@@ -16,14 +18,18 @@ function copyFolderSync(from, to) {
 }
 
 // Define la carpeta de origen y la carpeta de destino
-const sourceDir = path.join(__dirname, 'src', 'sass');
+const sourceDir = path.join(__dirname, 'src', 'styles', 'init');
 const targetDir = path.join(__dirname, '..', '..', 'src', 'sass');
 
-// Copia la carpeta de origen a la carpeta de destino
-try {
-    copyFolderSync(sourceDir, targetDir);
-} catch (error) {
-    console.error('Error al copiar la carpeta', error);
+// Verifica si la carpeta de destino ya existe
+if (!fs.existsSync(targetDir)) {
+    // Copia la carpeta de origen a la carpeta de destino
+    try {
+        copyFolderSync(sourceDir, targetDir);
+        console.log('Carpeta copiada');
+    } catch (error) {
+        console.error('Error al copiar la carpeta', error);
+    }
+} else {
+    console.log('La carpeta de destino ya existe. No se copian los archivos.');
 }
-
-console.log('Carpeta copiada');
